Remove GLHelper.lineWidth
git-svn-id: svn+ssh://solsvn.oma.be/home/sidc/ssr/SWHV_repos/trunk/client/jhv-3d-wcs@6373 b4e469a2-07ce-4b26-9273-4d7d95a670c7

+++ src/plugins/jhv-3d-plugin-pfss/src/org/helioviewer/jhv/plugins/pfssplugin/PfssSettings.java
@@ -11,7 +11,7 @@ public class PfssSettings {
 
     public static final float LINE_ALPHA = 1;
 
-    public static final double LINE_WIDTH = 0.5;
+    public static final float LINE_WIDTH = 0.5f;
 
     public static final int POINTS_PER_LINE = 40;
